+++ R/path.R
@@ -1,6 +1,13 @@
 
+`_renv_alpha` <- c(letters, LETTERS)
+
 renv_path_absolute <- function(path) {
-  grepl("^(?:[~/\\]|[a-zA-Z]:)", path)
+
+  substr(path, 1L, 1L) %in% c("~", "/", "\\") || (
+    substr(path, 1L, 1L) %in% `_renv_alpha` &&
+    substr(path, 2L, 3L) %in% c(":/", ":\\")
+  )
+
 }
 
 renv_path_within <- function(path, parent) {
@@ -41,8 +48,6 @@ renv_path_normalize_unix <- function(path,
 #
 # furthermore, it appears that shortPathName() can mis-encode its result for
 # strings marked with latin1 encoding?
-# https://github.com/rstudio/renv/issues/629
-#
 # https://github.com/rstudio/renv/issues/629
 renv_path_normalize_win32 <- function(path,
                                       winslash = "/",